Bootstrap popover trigger on hover instead of click

enhanced Bootstrap 3 popover by hovering, see bsplus::bs_embed_popover for details. Everything is the same but has additional trigger method, default "hover"

bsHoverPopover(
  tag,
  title = NULL,
  content = NULL,
  placement = "top",
  trigger = "hover",
  ...
)

Arguments

tag
htmltools::tag, generally htmltools::tags$button() or htmltools::tags$a(), into which to embed the popover

title
character, title for the popover, generally text

content
character, content for the popover body, can be HTML

placement
character, placement of the popover with respect to tag

trigger
trigger method, default "hover", one of click | hover | focus | manual.

...
other named arguments, passed to bsplus::bs_set_data()

Value

shiny element

Examples

if(interactive()){
    library(shiny)
    ui <- fluidPage(
        useSps(),
        actionButton('a', 'a') %>%
            bsHoverPopover(title = "title a", content = "a", placement = "bottom"),
        tags$a("b") %>%
            bsHoverPopover(title = "title b", content = "b", placement = "bottom")
    )
    server <- function(input, output, session) {}
    shinyApp(ui, server)
}