exploreDDS

Convenience wrapper function to transform raw read counts using the DESeq2 package transformations methods. The input file has to contain all the genes, not just differentially expressed ones.

exploreDDS(
  countMatrix,
  targets,
  cmp = cmp[[1]],
  preFilter = NULL,
  transformationMethod = "raw"
)

Arguments

countMatrix
date.frame or matrix containing raw read counts.

targets
targets data.frame.

cmp
character matrix where comparisons are defined in two columns. This matrix should be generated with the systemPipeR::readComp() function from the targets file. Values used for comparisons need to match those in the Factor column of the targets file.

preFilter
allows removing rows in which there are very few reads. Accepts a numeric value with the minimum of total reads to keep. Default is NULL.

transformationMethod
a character string indicating which transformation method it will be used on the raw read counts. Supported methods include rlog and vst using the DESeq2 package or default raw for no data transformation.

Value

returns an object of class DESeqTransform.

Details

Note that you can use the resulting transformed values in the transformationMethod argument only for visualization and clustering, not for differential expression analysis which needs raw counts. Users are strongly encouraged to consult the DESeq2 vignette for more detailed information on this topic and how to properly run DESeq2 on data sets with more complex experimental designs.

References

For more details on DESeq2, please consult the following page: http://bioconductor.org/packages/release/bioc/html/DESeq2.html For more details on targets file definition, please consult the following page: http://www.bioconductor.org/packages/release/bioc/vignettes/systemPipeR/inst/doc/systemPipeR.html#25_structure_of_targets_file

Examples

suppressPackageStartupMessages({library(systemPipeR)})

#> Warning: replacing previous import ‘Biostrings::pattern’ by ‘grid::pattern’ when loading ‘systemPipeR’

## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)
## Run
exploredds <- exploreDDS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="raw")

#> Warning: some variables in design formula are characters, converting to factors

exploredds

#> class: DESeqDataSet 
#> dim: 116 18 
#> metadata(1): version
#> assays(1): counts
#> rownames(116): AT1G01010 AT1G01020 ... ATMG00180 ATMG00200
#> rowData names(0):
#> colnames(18): M1A M1B ... V12A V12B
#> colData names(1): condition