exploreDSS Convenience wrapper function to transform raw read counts using the DESeq2 package transformations methods. The input file has to contain all the genes, not just differentially expressed ones.

exploreDSS Convenience wrapper function to transform raw read counts using the DESeq2 package transformations methods. The input file has to contain all the genes, not just differentially expressed ones.

exploreDSS(
  countMatrix,
  targets,
  cmp = cmp[[1]],
  preFilter = NULL,
  transformationMethod = "raw"
)

Arguments

countMatrix
date.frame or matrix containing raw read counts.

targets
targets data.frame.

cmp
character matrix where comparisons are defined in two columns. This matrix should be generated with the readComp() function from the targets file. Values used for comparisons need to match those in the Factor column of the targets file.

preFilter
allows removing rows in which there are very few reads. Accepts a numeric value with the minimum of total reads to keep. Default is NULL.

transformationMethod
a character string indicating which transformation method it will be used on the raw read counts. Supported methods include rlog and vst using the DESeq2 package or default raw for no data transformation.

Value

returns an object of class DESeqTransform.

Details

Note that you can use the resulting transformed values in the transformationMethod argument only for visualization and clustering, not for differential expression analysis which needs raw counts. Users are strongly encouraged to consult the DESeq2 vignette for more detailed information on this topic and how to properly run DESeq2 on data sets with more complex experimental designs.

References

For more details on DESeq2, please consult the following page: http://bioconductor.org/packages/release/bioc/html/DESeq2.html For more details on targets file definition, please consult the following page: http://www.bioconductor.org/packages/release/bioc/vignettes/systemPipeR/inst/doc/systemPipeR.html#25_structure_of_targets_file

Examples

library(systemPipeR)

#> Loading required package: Rsamtools

#> Loading required package: GenomeInfoDb

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: ‘IRanges’

#> The following objects are masked from ‘package:glue’:
#> 
#>     collapse, trim

#> Loading required package: GenomicRanges

#> Loading required package: Biostrings

#> Loading required package: XVector

#> 
#> Attaching package: ‘Biostrings’

#> The following object is masked from ‘package:base’:
#> 
#>     strsplit

#> Loading required package: ShortRead

#> Loading required package: BiocParallel

#> Loading required package: GenomicAlignments

#> Loading required package: SummarizedExperiment

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

#> 
#> Attaching package: ‘ShortRead’

#> The following object is masked from ‘package:magrittr’:
#> 
#>     functions

#> 

## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)
## Run
exploredss <- exploreDSS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="raw")

#> Warning: some variables in design formula are characters, converting to factors