Match height of one element to the other element

Match the height of one element to the second element. If the height of second element change, the height of first element will change automatically

heightMatcher(div1, div2, isID = TRUE)

Arguments

div1
element ID, or jquery selector if isID = FALSE. The first element that you want to match the height to the other element

div2
matched element ID or selector, the other element

isID
bool, if TRUE, div1 and div2 will be treated as ID, otherwise you can use complex jquery selector

Value

will be run as javascript

Examples

if(interactive()){
    library(shiny)

    ui <- fluidPage(
        useSps(),
        column(
            3, id = "a",
            style = "border: 1px black solid",
            p("a")
        ),
        shinyjqui::jqui_resizable(column(
            6, id ="b",
            style = "border: 1px black solid",
            p("b")
        )),
        heightMatcher("a", "b")
    )

    server <- function(input, output, session) {

    }
    # Try to drag `b` from bottom right corner and see what happens to `a`
    shinyApp(ui, server)
}