Create data receive methods for plot tabs

This function specify for each input data type in a plot tab, where the data is coming from, how to validate incoming data. To use this function, make sure there is a SPS project and config/tabs.csv exists.

makePlotData(
  dataset_id = "data",
  dataset_label = "Raw data",
  receive_datatab_ids = "data_template",
  vd_expr = spsValidate({     if (is.data.frame(mydata$data))          TRUE     else
    stop("Data xx needs to be a dataframe or tibble") }),
  app_path = getwd(),
  use_string = FALSE
)

Arguments

dataset_id
string, length 1, a unique ID within this plot tab.

dataset_label
string, length 1, what label to display on UI for this type of input data

receive_datatab_ids
a vector of tab IDs: For this kind of data input, which data tabs that can be used as input source(s). For example, if this plot tab requires a dataframe and can be produced from "data_df1" or "data_df2", receive_datatab_ids = c("data_df1", "data_df2"). These options are later rendered as a drop down menu for users to choose for this where they have prepared the required data from.

vd_expr
what expression to validate(check) the incoming data set. Usually it is a spsValidate object

app_path
SPS project folder

use_string
bool, if you don't want to parse vd_expr, use quoted string for vd_expr amd turn this to TRUE. see newTabPlot

Value

a special list storing the incoming data info

Details

For the validation expression, the incoming data is stored in a reactive values object, and you can access this data object by mydata$dataset_id, e.g. the dataset_id is "raw_data", then when the time you validate this type of incoming data set, a variable mydata$raw_data is accessible.

It is recommended to create data tabs first before running this function, because receive_datatab_ids required data tab id exists in the tabs.csv file.

Examples

spsInit(change_wd = FALSE, overwrite = TRUE)

#> [SPS-INFO] 2020-08-23 11:27:02 Start to create a new SPS project

#> Warning: [SPS-WARNING] 2020-08-23 11:27:02 Overwrite things in /home/dcassol/danielac@ucr.edu/testing/today/systemPipeShiny/docs/reference/SPS_20200823

#> [SPS-INFO] 2020-08-23 11:27:02 Now copy files
#> [SPS-INFO] 2020-08-23 11:27:02 Create SPS database
#> [SPS-INFO] 2020-08-23 11:27:02 Created SPS database method container
#> [SPS-INFO] 2020-08-23 11:27:02 Creating SPS db...
#> [SPS-DANGER] 2020-08-23 11:27:02 Db created at '/home/dcassol/danielac@ucr.edu/testing/today/systemPipeShiny/docs/reference/SPS_20200823/config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-08-23 11:27:02 Key md5 5598f5d80ec825b338f4d7ca55b3e8cf
#> [SPS-INFO] 2020-08-23 11:27:02 SPS project setup done!

newTabData("data_df1", "df 1",
           app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"),
           open_file = FALSE)

#> Error: [SPS-ERROR] 2020-08-23 11:27:02 File SPS_20200823/R/tab_vs_data_df1.R exists.

newTabData("data_df2", "df 2",
           app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"),
           open_file = FALSE)

#> Error: [SPS-ERROR] 2020-08-23 11:27:02 File SPS_20200823/R/tab_vs_data_df2.R exists.

plotdata_raw <- makePlotData("raw", "raw data",
             receive_datatab_ids = "data_df1",
             vd_expr = spsValidate({
                 if(!is.data.frame(mydata$raw))
                     stop("Input raw data need to be a dataframe")
             }, vd_name = "Validate raw data"),
             app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"))
plotdata_meta <- makePlotData("meta", "meta data",
             receive_datatab_ids = c("data_df1", "data_df2"),
             vd_expr = spsValidate({
                 if(!is.data.frame(mydata$meta))
                     stop("Input raw data need to be a dataframe")
                 if(nrow(mydata$meta) < 1)
                     stop("Input raw data need to have at least one row")
             }, vd_name = "Validate meta data"),
             app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"))
newTabPlot("plot_test1",
           app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"),
           plot_data = list(plotdata_raw, plotdata_meta))

#> Error: [SPS-ERROR] 2020-08-23 11:27:02 File SPS_20200823/R/tab_vs_plot_test1.R exists.