Create data tab preprocess methods

On a data tab, given the same uploaded data set, users can choose different ways to preprocess the data and therefore different preprocessing methods will lead to different plot tab options. Every call of this function defines a preprocess method: 1. data validation expression before preprocess; 2. actual preprocess expression; 3. plot options after preprocess.

makePrepro(
  method_id = "md1",
  label = "New method1",
  vd_expr = spsValidate(is.data.frame(data_filtered)),
  pre_expr = {     data_filtered },
  plot_options = "default",
  use_string = FALSE
)

Arguments

method_id
string, length 1, a unique ID within this data tab.

label
string, length 1, what label to display on UI for users to choose as a preprocess option

vd_expr
expression, usually a spsValidate object. Before preprocess if there is any additional validation that is special to this preprocess method, you can specify here

pre_expr
The actual preprocess expression. You should use a pre-created variable called data_filtered to start and this is the object that contains filtered data after users filtering on the UI. In the end of this expression, you should return a preprocessed dataframe or whatever object type that can be accepted by the desired plot tab. It is recommended to write the preprocess method into a function and directly call the function here, e.g.

myPreprocess <- function(data){
    if(is.numeric(data[ ,1]))
        data[ ,1] <- data[ ,1] + 1
    return(data)
}
makePrepro(
 ...,
 pre_expr = myPreprocess(data_filtered),
 ...
)

plot_options
plot tab IDs: if data is preprocessed by this method, what kind of plots it can make, specify plot tab IDs in a vector. Note: unlike the receive_datatab_ids argument in makePlotData that requires the config/tabs.csv exists, this argument doesn't require the config file or the plot tab to be existing. You can use any ID(s) here. The ID checking is postponed when the genGallery function runs on app start. "default" means all possible plot tabs, the same as type = 'plot' in genGallery.

use_string
same as the argument in newTabPlot, controls vd_expr and pre_expr in this function

Value

a special list that contains all info for a preprocess method

Details

for vd_expr, pre_expr a variable called data_filtered is accessible and is the object where data stored. One should use this object to do validation or preprocess. See examples.

Examples

spsInit(change_wd = FALSE, overwrite = TRUE, project_name = "SPS_prepro")

#> [SPS-INFO] 2020-09-29 17:09:56 Start to create a new SPS project

#> Warning: [SPS-WARNING] 2020-09-29 17:09:56 Overwrite things in /home/dcassol/danielac@ucr.edu/testing/systemPipeShiny/docs/reference/SPS_prepro

#> [SPS-INFO] 2020-09-29 17:09:56 Now copy files

#> [SPS-INFO] 2020-09-29 17:09:56 Create SPS database

#> [SPS-INFO] 2020-09-29 17:09:56 Created SPS database method container

#> [SPS-INFO] 2020-09-29 17:09:56 Creating SPS db...

#> [SPS-DANGER] 2020-09-29 17:09:56 Db created at '/home/dcassol/danielac@ucr.edu/testing/systemPipeShiny/docs/reference/SPS_prepro/config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-09-29 17:09:56 Key md5 efdd077f02f5d5318e8d4885489e4d50
#> [SPS-INFO] 2020-09-29 17:09:56 SPS project setup done!

# first preprocess method
prepro_log <- makePrepro(
    "log", "take log of first column",
    vd_expr = spsValidate({
        if(!is.data.frame(data_filtered))
            stop("Input input data need to be a dataframe")
    }, vd_name = "log method pre-checks"),
    pre_expr = {
        if(is.numeric(data_filtered[ ,1]))
            {if(all(data_filtered[ ,1] > 0)){
                data_filtered[ ,1] <- log(data_filtered[ ,1])}
            }
        data_filtered
    },
    plot_options = c("plot_xx1", "plot_xx2")
    )
## remember to save these helper functions in a R scripts under the R
## folder. They will be automatically sourced when app starts.
myPreprocess <- function(data){
    if(is.numeric(data[ ,1]))
        data[ ,1] <- data[ ,1] + 1
    return(data)
}
myVd <- function(data, vd_name){
    spsValidate({
        if(!is.data.frame(data))
            stop("Input input data need to be a dataframe")
    }, vd_name = vd_name)
}
# second preprocess method
prepro_addone <- makePrepro(
    "addone", "add one to first column",
    vd_expr = myVd(data_filtered, "add one method pre-checks"),
    pre_expr = myPreprocess(data_filtered),
    plot_options = c("plot_xx1")
)
# Combine two methods and make a new data tab
newTabData("data_test1",  "test 1",
           app_path = "SPS_prepro",
           prepro_methods = list(prepro_log, prepro_addone)
           )

#> Error: [SPS-ERROR] 2020-09-29 17:09:56 File SPS_prepro/R/tab_vs_data_test1.R exists.