SPS terminal message logging methods

If SPS use_crayonoption is TRUE, the message will be colorful. "INFO" level spawns message, "WARNING" is warning, "ERROR" spawns stop, other levels use cat. spsinfo, spswarn, spserror are higher level wrappers of msg. The only difference is they have SPS- prefix.

spsinfo has an additional arg verbose. This arg works similar to all other verbose args in this package, if not specified, it follows the project option, can be be forced to TRUE and FALSE. TRUE will forcefully generate the msg, and FALSE will be no message.

msg(
  msg,
  level = "INFO",
  .other_color = "white",
  info_text = "INFO",
  warning_text = "WARNING",
  error_text = "ERROR"
)

spsinfo(msg, verbose = NULL)

spswarn(msg)

spserror(msg)

Arguments

msg
a character string of message or a vector of character strings, each item in the vector presents one line of words

level
typically, one of "INFO", "WARNING", "ERROR", lower case OK. Other levels ok.

.other_color
hex color code or named colors, when levels are not in "INFO", "WARNING", "ERROR", this value will be used

info_text
info level text prefix

warning_text
warning level text prefix

error_text
error level text prefix

verbose
bool, default get from sps project options, can be overwritten

Value

see description

Details

If you want to use this function to generate colorful log messages but not in SPS framework. Use spsOption to set up the option on very top of your scripts: spsOption("use_crayon", TRUE).

Examples

msg("this is info")

#> [INFO] 2020-09-29 17:09:56 this is info

msg("this is warning", "warning")

#> Warning: [WARNING] 2020-09-29 17:09:56 this is warning

try(msg("this is error", "error"))

#> Error : [ERROR] 2020-09-29 17:09:56 this is error

msg("this is other", "error2")

#> [error2] 2020-09-29 17:09:56 this is other

spsinfo("some msg, verbose false", verbose = FALSE) # will not show up
spsinfo("some msg, verbose true", verbose = TRUE)

#> [SPS-INFO] 2020-09-29 17:09:56 some msg, verbose true

spswarn("sps warning")

#> Warning: [SPS-WARNING] 2020-09-29 17:09:56 sps warning

try(spserror("sps error"))

#> Error : [SPS-ERROR] 2020-09-29 17:09:56 sps error