SPS snapshots container

Initiate this container at the global level. This container is used to communicate plotting tabs with the canvas tab

Public fields

plot_ui

a list of plot UI snapshots will be stored here. You shouldn't manually edit this list, use add/getUI method

plot_server

a list of plot server snapshots will be stored here. You shouldn't manually edit this list, use add/getServer method

Methods

Public methods

plotContainer$new()

plotContainer$addUI()

plotContainer$getUI()

plotContainer$addServer()

plotContainer$getServer()

plotContainer$notifySnap()

plotContainer$clone()

Method new()

Usage

plotContainer$new()

Method addUI()

add plot UI to the container. use it to wrap around the original plot out put function

Usage

plotContainer$addUI(plot_DOM, tab_id)

Arguments

plot_DOM

Plotting output function, like plotOutput

tab_id

unique ID, usually use in a module and use the tab ID

Method getUI()

Get plot UI to the container.

Usage

plotContainer$getUI(tab_id, plot_id_new = NULL)

Arguments

tab_id

unique ID, usually the tab ID if used in a module

plot_id_new

string, usually if taking a snapshot of a plot DOM, they can't use the same ID on HTML. When get the UI, change the ID to a new value

Returns

a saved plotting DOM

Method addServer()

add plot server function to the container. use it to wrap around the original plot output server function, like renderPlot, renderPlotly

Usage

plotContainer$addServer(
  render_func,
  tab_id,
  expr,
  env = parent.frame(),
  quoted = FALSE,
  ...
)

Arguments

render_func

plot output server function

tab_id

unique ID, usually the tab ID if used in a module

expr

the reactive expression to render the plot, same as expr in other render function

env

default parent.frame(), see shiny documents

quoted

Is you expr quoted?

...

additional args to pass to the original render function

Method getServer()

Get plot UI to the container.

Usage

plotContainer$getServer(tab_id)

Arguments

tab_id

unique ID, usually the tab ID if used in a module

Returns

saved plot server function

Method notifySnap()

notify the snapshot tab a snapshot has been added

Usage

plotContainer$notifySnap(tab_id, skip = 1, reset = FALSE, set_to = NULL)

Arguments

tab_id

unique ID, usually the tab ID if used in a module

skip

positive integer, This function is usually bound to the plotting button. When clicked, the first number of skip clicks will not add the snapshot to the canvas tab.

reset

bool, to reset the count and return nothing

set_to

positive integer, reset the count of a tab to be a certain number

Returns

a vector of tab_id and count number as a character string

Method clone()

The objects of this class are cloneable with this method.

Usage

plotContainer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if(interactive()){
    library(shiny)
    library(shinydashboard)
    library(shinyjs)
    plots = plotContainer$new()
    mod1_UI <- function(id) {
        ns <- NS(id)
        tagList(
            h1("a tiny example of how Canvas work in SPS"),
            actionButton(ns("render"), "render"),
            jqui_resizable(plots$addUI(plotlyOutput(ns("plot1")), id)),
            sliderInput(ns("slide"), label = "rows",
                        min = 1, max = nrow(iris),
                        value = nrow(iris))

        )
    }
    mod1 <- function(input, output, session, shared) {
        observeEvent(input$render, {
            output$plot1 <- plots$addServer(renderPlotly, 'mod1', {
                ggplotly(ggplot(iris[1:input$slide, ],
                                aes(Sepal.Length, Sepal.Width)) +
                             geom_point(aes(colour = Species)))
            })
            shared$snap_signal <- plots$notifySnap("mod1")
            req(shared$snap_signal)
            toastr_info(
                glue("Snapshot {glue_collapse(shared$snap_signal, '-')}",
                     "added to canvas"),
                        position = "bottom-right")
        })
    }
    mod2_UI <- function(id){
        ns <- NS(id)
        tagList(
            actionButton(ns("refresh"), 'refresh'),
            sliderTextInput(
                inputId = ns("ncols"),
                label = "Number of columns per row to initiate canvas:",
                choices = c(1:4, 12), selected = 2, grid = TRUE
            ),
            fluidRow(uiOutput(ns("new")), class = "sps-canvas")
        )
    }

    mod2 <- function(input, output, session, shared) {
        ns <- session$ns
        make_plots <- reactiveValues(ui = list(), server = list())
        observeEvent(shared$snap_signal, {
            tab_id <- shared$snap_signal[1]
            new_plot_id <- glue("{tab_id}-{shared$snap_signal[2]}")
            print(new_plot_id)
            make_plots$ui[[new_plot_id]] <-
                list(plots$getUI(tab_id, ns(new_plot_id)), ns(new_plot_id))
            make_plots$server[[new_plot_id]] <- plots$getServer(tab_id)
        })
        observeEvent(input$refresh, {
            ui <- make_plots$ui
            output$new <- renderUI({
                sapply(seq_along(ui), function(i){
                    column(
                        width = 12/isolate(input$ncols),
                        class = "collapse in",
                        id = glue("{ui[[i]][2]}-container"),
                        div(class = "snap-drag bg-primary",
                            h4(glue("Plot {ui[[i]][2]}")),
                            tags$button(
                                class = "btn action-button canvas-close",
                                icon("times"),
                                `data-toggle`="collapse",
                                `data-target`=glue("#{ui[[i]][2]}-container"),
                                `plot-toggle` = ui[[i]][2]
                            )
                        ),
                        jqui_resizable(make_plots$ui[[i]][[1]]),
                        tags$script(glue(.open = '@', .close = '@',
                                         '$("#@ui[[i]][2]@-container")',
                                         '.draggable({ handle: ".snap-drag"})'))
                    )
                }, simplify = FALSE) %>%{
                    fluidRow(id = ns('plots'),
                             tags$head(tags$style(
                                 '
                             .snap-drag {
                             opacity: 0;
                             };'),
                                 tags$style(
                                     '
                             .snap-drag button{
                             position: absolute;
                             outline: none;
                             background-color: Transparent;
                             top: 2px;
                             right: 4px;
                             };'),
                                 tags$style(
                                     '
                             .snap-drag h4{
                             margin-bottom: 0;
                             padding-bottom: 10px;
                             };'),
                                 tags$style(
                                     '
                             .snap-drag button:focus {
                             outline: 0 !important;
                             box-shadow: none !important;
                             };'),
                                 tags$style('.snap-drag:hover {
                             opacity: 1;
                             }
                             '
                                 )),
                             tagList(.)
                    )
                }
            })
            for(plot_id in names(make_plots$server)){
                output[[plot_id]] <- make_plots$server[[plot_id]]
            }
        })
        observeEvent(input$hide_title, {
            shinyjs::toggleClass(selector = ".snap-drag", class = "collapse")
        })

    }
    ui <- dashboardPagePlus(
        header = dashboardHeaderPlus(),
        sidebar = dashboardSidebar(),
        body = dashboardBody(
            useShinyjs(),
            useSps(),
            mod1_UI("mod1"),
            mod2_UI("mod2")),
        title = "Test",
    )
    server <- function(input, output, session) {
        shared = reactiveValues()
        callModule(mod1, "mod1",  shared)
        callModule(mod2, "mod2",  shared)
    }
    shinyApp(ui, server)
}