Catch error, warning, message text

Catch error, warning, message by a toastr bar on shiny front end also log the text on backend console. Will return original value if not blocking at "warning" "message" level, and return NULL at error level. If blocks at error, function will be stopped and other code in the same reactive context will be blocked. If blocks at warning level, warning and error will be blocked; message level blocks all 3 levels. The blocking works similar to shiny's req() and validate(). If anything inside fails, it will block the rest of the code in your reactive expression domain

shinyCatch(
  expr,
  position = "bottom-right",
  blocking_level = "none",
  shiny = TRUE
)

Arguments

expr
expression

position
toastr position, one of: c("top-right", "top-center", "top-left", "top-full-width", "bottom-right", "bottom-center", "bottom-left", "bottom-full-width")

blocking_level
what level you want to block the execution, one of "error", "warning", "message"

shiny
bool, It is also possible to use without a shiny session. Only shows on console log, works very similar as tryCatch and can block at multiple levels

Value

see description

Examples

if(interactive()){
    ui <- fluidPage(
        useToastr(),
        actionButton("btn1","error and blocking"),
        actionButton("btn2","error no blocking"),
        actionButton("btn3","warning but still returns value"),
        actionButton("btn4","warning but blocking returns"),
        actionButton("btn5","message"),

        textOutput("text")
    )
    server <- function(input, output, session) {
        fn_warning <- function() {
            warning("this is a warning!")
            return(1)
        }
        observeEvent(input$btn1, {
            shinyCatch(stop("error with blocking"), blocking_level = "error")
            print("You shouldn't see me")
        })
        observeEvent(input$btn2, {
            shinyCatch(stop("error without blocking"))
            print("I am not blocked by error")
        })
        observeEvent(input$btn3, {
            return_value <- shinyCatch(fn_warning())
            print(return_value)
        })
        observeEvent(input$btn4, {
            return_value <- shinyCatch(fn_warning(), blocking_level = "warning")
            print(return_value)
            print("other things")
        })
        observeEvent(input$btn5, {
            shinyCatch(message("message"))
        })
    }
    shinyApp(ui, server)
}
#outside shiny examples
shinyCatch(message("this message"), shiny = FALSE)

#> [SPS-INFO] 2020-08-10 16:04:24 this message
#> 

try({shinyCatch(stop("this error"), shiny = FALSE); "no block"}, silent = TRUE)

#> [SPS-ERROR] 2020-08-10 16:04:24 this error

#> [1] "no block"

try({shinyCatch(stop("this error"), shiny = FALSE, blocking_level = "error")
    "blocked"},
    silent = TRUE)

#> [SPS-ERROR] 2020-08-10 16:04:24 this error