SPS account management functions

Initiate this container at global level. Methods in this class can help admins to manage accounts in a SPS project.

It uses a SQLite database, by default is created inside config directory on SPS initialization.

You can use it to add/remove users, change user roles, change password, match/verify account, password, role.

A default user account "user", with password "user", and a default admin account "admin" with password "admin" are create for you.

For app deployment, PLEASE create your own accounts and DELETE the default ones.

Super classes

systemPipeShiny::spsDb -> systemPipeShiny::spsEncryption -> spsaccount

Methods

Public methods

spsAccount$new()

spsAccount$accList()

spsAccount$accAdd()

spsAccount$accRemove()

spsAccount$accPassChange()

spsAccount$accRoleChange()

spsAccount$accMatch()

spsAccount$clone()

Inherited methods

systemPipeShiny::spsDb$createDb()

systemPipeShiny::spsDb$queryDel()

systemPipeShiny::spsDb$queryInsert()

systemPipeShiny::spsDb$queryUpdate()

systemPipeShiny::spsDb$queryValue()

systemPipeShiny::spsDb$queryValueDp()

systemPipeShiny::spsEncryption$decrypt()

systemPipeShiny::spsEncryption$encrypt()

systemPipeShiny::spsEncryption$keyChange()

systemPipeShiny::spsEncryption$keyGet()

Method new()

initialize a new SPS account container

Usage

spsAccount$new()

Method accList()

list all accounts of the app. Returns a dataframe

Usage

spsAccount$accList(include_pass = FALSE, db_name = "config/sps.db")

Arguments

include_pass

bool, include password hash column?

db_name

SPS database path

Method accAdd()

add an account to use the app

Usage

spsAccount$accAdd(acc_name, acc_pass, role = "user", db_name = "config/sps.db")

Arguments

acc_name

string, account name

acc_pass

string, account password

role

string, what kind role is this user, one of "user", "admin"

db_name

SPS database path

Method accRemove()

remove an account

Usage

spsAccount$accRemove(acc_name, db_name = "config/sps.db")

Arguments

acc_name

string, account name

db_name

SPS database path

Method accPassChange()

change password of an account

Usage

spsAccount$accPassChange(acc_name, acc_pass, db_name = "config/sps.db")

Arguments

acc_name

string, account name

acc_pass

string, account new password

db_name

SPS database path

Method accRoleChange()

change the role of an account

Usage

spsAccount$accRoleChange(acc_name, role, db_name = "config/sps.db")

Arguments

acc_name

string, account name

role

string, one of "user" or "admin"

db_name

SPS database path

Method accMatch()

Try to see if the account name exists and has the right password and role type, useful for login authentification.

Usage

spsAccount$accMatch(
  acc_name,
  acc_pass,
  role = "user",
  match_role = FALSE,
  db_name = "config/sps.db"
)

Arguments

acc_name

string, account name

acc_pass

string, account new password

role

string, one of "user" or "admin"

match_role

bool, also verify the account role type?

db_name

SPS database path

Method clone()

The objects of this class are cloneable with this method.

Usage

spsAccount$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

dir.create("config", showWarnings = FALSE)
spsOption("verbose", TRUE)
spsOption("use_crayon", TRUE)
# create a new container
db <- spsAccount$new()

#> [SPS-INFO] 2021-03-30 02:32:24 Created SPS account method container

#> [SPS-INFO] 2021-03-30 02:32:24 This container inherits all functions from spsEncryption and spsDb class

#> [SPS-INFO] 2021-03-30 02:32:24 Db connected

#> [SPS-INFO] 2021-03-30 02:32:24 Connected, but tables missing, seems like a new db.

db$createDb()

#> [SPS-INFO] 2021-03-30 02:32:24 Db connected

#> [SPS-INFO] 2021-03-30 02:32:24 Creating SPS db...

#> [SPS-INFO] 2021-03-30 02:32:24 Db write the meta table

#> [SPS-INFO] 2021-03-30 02:32:25 Db write the raw table

#> [SPS-INFO] 2021-03-30 02:32:25 Key generated and stored in db

#> [SPS-INFO] 2021-03-30 02:32:25 Db create admin account

#> [SPS-DANGER] 2021-03-30 02:32:25 Done, Db created at 'config/sps.db'. DO NOT share this file with others or upload to open access domains.
#> [SPS-INFO] 2021-03-30 02:32:25 Key md5 1063d421ee4afa11f59107755bf3ce0b

# list all accounts
db$accList()

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#>   account  role
#> 1   admin admin
#> 2    user  user

# add a new user
db$accAdd('user2', '!admin12345')

#> [SPS-INFO] 2021-03-30 02:32:25 Checking account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Inerted 1 rows

#> [SPS-INFO] 2021-03-30 02:32:25 Account user2 created.

# list all accounts include password hash
db$accList(include_pass = TRUE)

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#>   account                                                             pass
#> 1   admin 77eeb1d9862411f3eff3e6df741fac99d5484d1f01e6a0a5b5ec44647351016b
#> 2    user 7f1a841faa2f94803b6fd51472748754ad9126a7aae8ecb9d8ad8ae46fd6d769
#> 3   user2 6f487e08db8781d8bf9f5bb0bd8820b30ef50220aa8eefe7943432548fe8421b
#>    role
#> 1 admin
#> 2  user
#> 3  user

# change password of an account
db$accPassChange("user2", "$aaaaaaa")

#> [SPS-INFO] 2021-03-30 02:32:25 Checking account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> Error in db$accPassChange("user2", "$aaaaaaa"): object 'db' not found

# check if pass changed
db$accList(include_pass = TRUE)

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#>   account                                                             pass
#> 1   admin 77eeb1d9862411f3eff3e6df741fac99d5484d1f01e6a0a5b5ec44647351016b
#> 2    user 7f1a841faa2f94803b6fd51472748754ad9126a7aae8ecb9d8ad8ae46fd6d769
#> 3   user2 6f487e08db8781d8bf9f5bb0bd8820b30ef50220aa8eefe7943432548fe8421b
#>    role
#> 1 admin
#> 2  user
#> 3  user

# change the role of from user to admin
db$accRoleChange("user2", "admin")

#> [SPS-INFO] 2021-03-30 02:32:25 Checking account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> Error in db$accRoleChange("user2", "admin"): object 'db' not found

# check role change
db$accList()

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#>   account  role
#> 1   admin admin
#> 2    user  user
#> 3   user2  user

# remove a user
db$accRemove("user2")

#> [SPS-INFO] 2021-03-30 02:32:25 Checking account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Deleted 1 rows

#> [SPS-INFO] 2021-03-30 02:32:25 Account user2 removed

# check accounts again
db$accList()

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#>   account  role
#> 1   admin admin
#> 2    user  user

# check if username and password matches
db$accMatch(acc_name = "user", acc_pass = "user")

#> [SPS-INFO] 2021-03-30 02:32:25 Matching account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> [1] TRUE

# wrong pass
db$accMatch("user", "user123")

#> [SPS-INFO] 2021-03-30 02:32:25 Matching account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> [1] FALSE

# also check if the user has the right role
db$accMatch("user", "user", role = "user", match_role = TRUE)

#> [SPS-INFO] 2021-03-30 02:32:25 Matching account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> [1] TRUE

db$accMatch("user", "user", role = "admin", match_role = TRUE)

#> [SPS-INFO] 2021-03-30 02:32:25 Matching account info

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 Db connected

#> [SPS-INFO] 2021-03-30 02:32:25 Query sent

#> [SPS-INFO] 2021-03-30 02:32:25 OpenSSL key found md51063d421ee4afa11f59107755bf3ce0b

#> [1] FALSE