SPS database functions

Initiate this container at global level. Methods in this class can help admin to manage general information of sps. For now it only stores some meta data and the encryption key pairs. You can use this database to store other useful things, like user pass hash, IP, browsing info ...

A SQLite database by default is created inside config directory. If not, you can use createDb method to create one. On initiation, this class checks if the default db is there and gives warnings if not.

Methods

Public methods

spsDb$new()

spsDb$createDb()

spsDb$queryValue()

spsDb$queryValueDp()

spsDb$queryUpdate()

spsDb$queryDel()

spsDb$queryInsert()

spsDb$clone()

Method new()

Usage

spsDb$new()

Method createDb()

Create a snap hash database

Usage

spsDb$createDb(db_name = "config/sps.db")

Arguments

db_name

database path, you need to manually create parent directory if not exists

Method queryValue()

Query database

Usage

spsDb$queryValue(table, SELECT = "*", WHERE = "1", db_name = "config/sps.db")

Arguments

table

table name

SELECT

SQL select grammar

WHERE

SQL select where

db_name

database path

Returns

query result, usually a dataframe

Method queryValueDp()

Query database with dplyr grammar

Only supports simple selections, like comparison, %in%, between(), is.na(), etc. Advanced selections like wildcard, using outside dplyr functions like str_detect, grepl are not supported.

Usage

spsDb$queryValueDp(
  table,
  dp_expr = "select(., everything())",
  db_name = "config/sps.db"
)

Arguments

table

table name

dp_expr

dplyr chained expression, must use '.' in first component of the chain expression

db_name

database path

Returns

query result, usually a tibble

Method queryUpdate()

update(modify) the value in db

Usage

spsDb$queryUpdate(table, value, col, WHERE = "1", db_name = "config/sps.db")

Arguments

table

table name

value

new value

col

which column

WHERE

SQL where statement, conditions to select rows

db_name

database path

Method queryDel()

delete value in db

Usage

spsDb$queryDel(table, WHERE = "1", db_name = "config/sps.db")

Arguments

table

table name

WHERE

SQL where statement, conditions to select rows

db_name

database path

Method queryInsert()

Insert value to db

Usage

spsDb$queryInsert(table, value, db_name = "config/sps.db")

Arguments

table

table name

value

new values for the entire row

db_name

database path

WHERE

SQL where statement, conditions to select rows

Method clone()

The objects of this class are cloneable with this method.

Usage

spsDb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

dir.create("config")
mydb <- spsDb$new()

#> [SPS-INFO] 2020-09-11 01:18:49 Created SPS database method container

#> [SPS-INFO] 2020-09-11 01:18:49 Connected, but tables missing, seems like a new db.

#> [SPS-INFO] 2020-09-11 01:18:49 Use CreateDb method to create tables

mydb$createDb()

#> [SPS-INFO] 2020-09-11 01:18:49 Creating SPS db...

#> [SPS-DANGER] 2020-09-11 01:18:49 Db created at 'config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-09-11 01:18:49 Key md5 27b6603025bac37c72bd99352f53634d

mydb$queryValue("sps_meta")

#>            info          value
#> 1 creation_date 20200911011849

mydb$queryInsert("sps_meta", value = "'new1', '1'")

#> [SPS-INFO] 2020-09-11 01:18:49 Inerted 1 rows

mydb$queryValue("sps_meta")

#>            info          value
#> 1 creation_date 20200911011849
#> 2          new1              1

mydb$queryInsert("sps_meta", value = c("'new2'", "'2'"))

#> [SPS-INFO] 2020-09-11 01:18:49 Inerted 1 rows

mydb$queryValue("sps_meta")

#>            info          value
#> 1 creation_date 20200911011849
#> 2          new1              1
#> 3          new2              2

mydb$queryUpdate("sps_meta", value = '234',
                 col = "value", WHERE = "info = 'new1'")

#> [SPS-INFO] 2020-09-11 01:18:49 Updated 1 rows

mydb$queryValue("sps_meta")

#>            info          value
#> 1 creation_date 20200911011849
#> 2          new1            234
#> 3          new2              2

if (FALSE) {
    library(dplyr)
    mydb$queryValueDp(
        "sps_meta",
        dp_expr="filter(., info %in% c('new1', 'new2') %>% select(2)")
}
mydb$queryDel("sps_meta", WHERE = "value = '234'")

#> [SPS-INFO] 2020-09-11 01:18:49 Deleted 1 rows