SPS encryption functions

Methods in this class can help admin to encrypt files been output from sps. For now it is only used to encypt and decrypt snapshots. This class requires the SPS database. This class inherits all functions from the spsDb class, so there is no need to initiate the spsDb container.

This class is required to run a SPS app. This class needs to be initialized global level. This has already been written in global.R for you.

Super class

systemPipeShiny::spsDb -> spsencrypt

Methods

Public methods

spsEncryption$new()

spsEncryption$keyChange()

spsEncryption$keyGet()

spsEncryption$encrypt()

spsEncryption$decrypt()

spsEncryption$clone()

Inherited methods

systemPipeShiny::spsDb$createDb()

systemPipeShiny::spsDb$queryDel()

systemPipeShiny::spsDb$queryInsert()

systemPipeShiny::spsDb$queryUpdate()

systemPipeShiny::spsDb$queryValue()

systemPipeShiny::spsDb$queryValueDp()

Method new()

initialize a new class container

Usage

spsEncryption$new()

Method keyChange()

Change encryption key of a SPS project

Usage

spsEncryption$keyChange(db_name = "config/sps.db")

Arguments

db_name

database path

Method keyGet()

Get encryption key from db of a SPS project

Usage

spsEncryption$keyGet(db_name = "config/sps.db")

Arguments

db_name

database path

Method encrypt()

Encrypt raw data or a file with key from a SPS project

Usage

spsEncryption$encrypt(
  data,
  out_path = NULL,
  overwrite = FALSE,
  db_name = "config/sps.db"
)

Arguments

data

raw vector or a file path

out_path

if provided, encrypted data will be write to a file

overwrite

if out_path file exists, overwrite?

db_name

database path

Method decrypt()

Decrypt raw data or a file with key from a SPS project

Usage

spsEncryption$decrypt(
  data,
  out_path = NULL,
  overwrite = FALSE,
  db_name = "config/sps.db"
)

Arguments

data

raw vector or a file path

out_path

if provided, encrypted data will be write to a file

overwrite

if out_path file exists, overwrite?

db_name

database path

Method clone()

The objects of this class are cloneable with this method.

Usage

spsEncryption$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

dir.create("config", showWarnings = FALSE)
spsOption('verbose', TRUE)
my_ecpt <- spsEncryption$new()

#> [SPS-INFO] 2020-09-29 17:10:07 Created SPS encryption method container

#> [SPS-INFO] 2020-09-29 17:10:07 This container inherits all functions from spsDb class

#> [SPS-INFO] 2020-09-29 17:10:07 Db connected

#> [SPS-INFO] 2020-09-29 17:10:07 Default SPS-db found and is working

my_ecpt$createDb()

#> [SPS-INFO] 2020-09-29 17:10:07 Db connected

#> [SPS-INFO] 2020-09-29 17:10:07 Creating SPS db...

#> [SPS-INFO] 2020-09-29 17:10:07 Db write meta

#> [SPS-INFO] 2020-09-29 17:10:07 Key generated and stored in db

#> [SPS-DANGER] 2020-09-29 17:10:07 Db created at 'config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-09-29 17:10:07 Key md5 9831b956a1779597d63b451992e81fed

my_ecpt$keyChange()

#> [SPS-INFO] 2020-09-29 17:10:07 Db connected

#> [SPS-INFO] 2020-09-29 17:10:07 You new key has been generated and saved in db

#> [SPS-INFO] 2020-09-29 17:10:07 md5 1997ad6a090e91bbfd04008af030c528

# imagine a file has one line "test"
writeLines(text = "test", con = "test.txt")
# encrypt the file
my_ecpt$encrypt("test.txt", "test.bin", overwrite = TRUE)

#> [SPS-INFO] 2020-09-29 17:10:07 Db connected

#> [SPS-INFO] 2020-09-29 17:10:07 Query sent

#> [SPS-INFO] 2020-09-29 17:10:07 OpenSSL key found md51997ad6a090e91bbfd04008af030c528

#> [SPS-INFO] 2020-09-29 17:10:07 Data encrypted

#> [SPS-INFO] 2020-09-29 17:10:07 File write to /home/dcassol/danielac@ucr.edu/testing/systemPipeShiny/docs/reference/test.bin

# decrypt the file
my_ecpt$decrypt("test.bin", "test_decpt.txt", overwrite = TRUE)

#> [SPS-INFO] 2020-09-29 17:10:07 File read into memory

#> [SPS-INFO] 2020-09-29 17:10:07 Db connected

#> [SPS-INFO] 2020-09-29 17:10:07 Query sent

#> [SPS-INFO] 2020-09-29 17:10:07 OpenSSL key found md51997ad6a090e91bbfd04008af030c528

#> [SPS-INFO] 2020-09-29 17:10:07 Data decrypted

#> [SPS-INFO] 2020-09-29 17:10:07 File write to /home/dcassol/danielac@ucr.edu/testing/systemPipeShiny/docs/reference/test_decpt.txt

# check the decrypted file content
readLines('test_decpt.txt')

#> [1] "test"