Easy and simple UI and server for a SPS custom tab

SPS custom tab simple UI and server , spsEzUI must use together with the spsEzServer function. The easiest way to use is to use spsNewTab function to create both.

spsEzUI(
  desc = "",
  tab_title = "Tab Title",
  plot_title = "My Plot",
  plot_control = shiny::tagList()
)

spsEzServer(
  plot_code,
  example_data_path = system.file(package = "systemPipeShiny", "app", "data",
    "iris.csv"),
  other_server_code = ""
)

Arguments

desc
character string, length 1 in markdown format. Tab description and instructions. You can make type it in multiple lines but in only one string (one pair of quotes). e.g.

"
# some desc
## second line,
- bullet 1
- bullet 2
"

tab_title
string, tab title

plot_title
string, plot title

plot_control
some Shiny components (UI) to control the plot, like plot title, x,y labels, color, font size, etc. Group all components in a shiny tagList.

plot_code
some R code to make the plot.

example_data_path
example dataset path, this dataset will be loaded on app start to display

other_server_code
optional, other server R code to run for this tab

Value

spsEzUI returns a shiny module UI function, spsEzServer returns the server function

See also

Examples

# use `spsInit()` to create an SPS project and use `spsNewTab("Your_tabID", template = "easy")`
# to create a new tab file. The specified use of these two functions is in that file.