Print SPS options

Make sure you have created the app directory and it has config/config.yaml file.

spsOptDefaults prints out all default and other avaliable values for each option. spsOptions print all current set option values.

Note: the spsUtil::spsOption is used to get or set a single option value. spsOptions is used to print all current option values. If you need to set all values at once, use the global.R file under SPS project root.

spsOptDefaults(app_path = getwd())

spsOptions(app_path = getwd(), show_legend = TRUE)

Arguments

app_path
path, where is the app directory

show_legend
bool, show the color legend?

Value

cat to console SPS option values

Examples

if(interactive()){
    # start a SPS project
    spsInit(open_files = FALSE)
    viewSpsDefaults()
    # change a few options
    options(sps = list(
        mode = "server",
        warning_toast = TRUE,
        login_screen = FALSE,
        login_theme = "vhelix",
        use_crayon = TRUE
    ))
    # view current options
    spsOptions()
}