View SPS project 'config/tabs.csv' information

View SPS project 'config/tabs.csv' information

spsTabInfo(return_type = "print", n_print = 40, app_path = getwd())

Arguments

return_type
one of 'print', 'data', 'colnames', or a specified column name

n_print
how many lines of tab info you want to print out

app_path
SPS project root

Value

return depends on return_type

Details

'print' will print out the entire tabs.csv, you can specify n_print for how many lines you want to print;

'data' will return the tab info tibble

'colnames' will return all column names of tab info file

A column name will extract the specified column out and return as a vector

Examples

spsInit(project_name = "SPS_tabinfo", overwrite = TRUE,
        change_wd = FALSE, open_files = FALSE)

#> [SPS-INFO] 2021-09-18 02:42:33 Start to create a new SPS project

#> [SPS-INFO] 2021-09-18 02:42:33 Create project under /home/runner/work/systemPipeShiny/systemPipeShiny/docs/reference/SPS_tabinfo

#> [SPS-INFO] 2021-09-18 02:42:33 Now copy files

#> [SPS-INFO] 2021-09-18 02:42:33 Create SPS database

#> [SPS-INFO] 2021-09-18 02:42:33 Created SPS database method container

#> [SPS-INFO] 2021-09-18 02:42:33 Creating SPS db...

#> [SPS-DANGER] 2021-09-18 02:42:33 Done, Db created at '/home/runner/work/systemPipeShiny/systemPipeShiny/docs/reference/SPS_tabinfo/config/sps.db'. DO NOT share this file with others or upload to open access domains.
#> [SPS-INFO] 2021-09-18 02:42:33 Key md5 228ec70ce15415db80115d61f9aa9199
#> [SPS-INFO] 2021-09-18 02:42:33 SPS project setup done!

# all lines
spsTabInfo("print", app_path = "SPS_tabinfo")

#> [SPS-INFO] 2021-09-18 02:42:33 Now check the tab info in tabs.csv 

#> [SPS-INFO] 2021-09-18 02:42:33 tab.csv info check pass

#> # A tibble: 11 × 8
#>    tab_id   display_label    type  type_sub image displayed tab_file_name plugin
#>    <chr>    <chr>            <chr> <chr>    <chr> <chr>     <chr>         <chr> 
#>  1 core_ab… About this App   core  ""       ""    1         No file for … "core"
#>  2 core_ca… Canvas           core  ""       ""    1         No file for … "core"
#>  3 core_we… Home             core  ""       ""    1         No file for … "core"
#>  4 core_ri… Internal use on… core  ""       ""    1         No file for … "core"
#>  5 core_top Top push bars    core  ""       ""    1         No file for … "core"
#>  6 module_… Module Main Page core  ""       ""    1         No file for … "core"
#>  7 wf       Workflow module  modu… ""       ""    1         No file for … "core"
#>  8 vs_rnas… RNAseq module    modu… ""       ""    1         No file for … "core"
#>  9 vs_esq   Quick ggplot mo… modu… ""       ""    1         No file for … "core"
#> 10 vs_main  custom tabs mai… core  ""       ""    1         No file for … "core"
#> 11 vs_exam… My custom plott… vs    "plot"   ""    1         tab_vs_examp… ""    

# 5 lines
spsTabInfo("print", app_path = "SPS_tabinfo", n_print = 5L)

#> [SPS-INFO] 2021-09-18 02:42:33 Now check the tab info in tabs.csv 

#> [SPS-INFO] 2021-09-18 02:42:33 tab.csv info check pass

#> # A tibble: 11 × 8
#>   tab_id       display_label type  type_sub image displayed tab_file_name plugin
#>   <chr>        <chr>         <chr> <chr>    <chr> <chr>     <chr>         <chr> 
#> 1 core_about   About this A… core  ""       ""    1         No file for … core  
#> 2 core_canvas  Canvas        core  ""       ""    1         No file for … core  
#> 3 core_welcome Home          core  ""       ""    1         No file for … core  
#> 4 core_right   Internal use… core  ""       ""    1         No file for … core  
#> 5 core_top     Top push bars core  ""       ""    1         No file for … core  
#> # … with 6 more rows

spsTabInfo("data", app_path = "SPS_tabinfo")

#> [SPS-INFO] 2021-09-18 02:42:33 Now check the tab info in tabs.csv 

#> [SPS-INFO] 2021-09-18 02:42:33 tab.csv info check pass

#> # A tibble: 11 × 8
#>    tab_id   display_label    type  type_sub image displayed tab_file_name plugin
#>    <chr>    <chr>            <chr> <chr>    <chr> <chr>     <chr>         <chr> 
#>  1 core_ab… About this App   core  ""       ""    1         No file for … "core"
#>  2 core_ca… Canvas           core  ""       ""    1         No file for … "core"
#>  3 core_we… Home             core  ""       ""    1         No file for … "core"
#>  4 core_ri… Internal use on… core  ""       ""    1         No file for … "core"
#>  5 core_top Top push bars    core  ""       ""    1         No file for … "core"
#>  6 module_… Module Main Page core  ""       ""    1         No file for … "core"
#>  7 wf       Workflow module  modu… ""       ""    1         No file for … "core"
#>  8 vs_rnas… RNAseq module    modu… ""       ""    1         No file for … "core"
#>  9 vs_esq   Quick ggplot mo… modu… ""       ""    1         No file for … "core"
#> 10 vs_main  custom tabs mai… core  ""       ""    1         No file for … "core"
#> 11 vs_exam… My custom plott… vs    "plot"   ""    1         tab_vs_examp… ""    

spsTabInfo("colnames", app_path = "SPS_tabinfo")

#> [SPS-INFO] 2021-09-18 02:42:33 Now check the tab info in tabs.csv 

#> [SPS-INFO] 2021-09-18 02:42:33 tab.csv info check pass

#> [1] "tab_id"        "display_label" "type"          "type_sub"     
#> [5] "image"         "displayed"     "tab_file_name" "plugin"       

spsTabInfo("tab_id", app_path = "SPS_tabinfo")

#> [SPS-INFO] 2021-09-18 02:42:33 Now check the tab info in tabs.csv 

#> [SPS-INFO] 2021-09-18 02:42:33 tab.csv info check pass

#>  [1] "core_about"   "core_canvas"  "core_welcome" "core_right"   "core_top"    
#>  [6] "module_main"  "wf"           "vs_rnaseq"    "vs_esq"       "vs_main"     
#> [11] "vs_example"