Validate expressions

this function is usually used on server side to validate input dataframe or some expression

spsValidate(
  expr,
  vd_name = "validation",
  pass_msg = glue("{vd_name} passed"),
  fail_msg = glue("{vd_name} failed"),
  shiny = TRUE,
  verbose = spsOption("verbose")
)

Arguments

expr
the expression to validate data or other things. It should return TRUE if pass or use stop("your message") if fail. Other types of return are acceptable but not recommended. As long as it is not empty or FALSE by the emptyIsFalse() function, it will return TRUE

vd_name
validate title

pass_msg
string, if pass, what message do you want to show

fail_msg
string, optional, if your expression does not contain the use of stop() for failure and only returns FALSE or other empty values, this message will show and generate shiny reactive stop

shiny
you can use this function outside shiny, see shinyCatch() for more

verbose
bool, show pass message? Default follows project verbose setting

Value

If expression returns empty or FALSE make it shiny reactive stop and no final return, else TRUE.

See also

Examples

spsOption("verbose", TRUE)
if(interactive()){
    ui <- fluidPage(
        useSps(),
        actionButton("vd1", "validate1"),
        actionButton("vd2", "validate2")
    )
    server <- function(input, output, session) {
        mydata <- datasets::iris
        observeEvent(input$vd1, {
            spsValidate({
                is.data.frame(mydata)
            }, vd_name = "Is df")
            print("continue other things")
        })
        observeEvent(input$vd2, {
            spsValidate({
                nrow(mydata) > 200
            }, vd_name = "more than 200 rows")
            print("other things blocked")
        })
    }
    shinyApp(ui, server)
}
# outside shiny example
mydata2 <- list(a = 1, b = 2)
spsValidate({(mydata2)}, "Not empty", shiny = FALSE)

#> [SPS-INFO] 2020-08-10 09:51:58 Not empty passed

#> [1] TRUE

try(spsValidate(is.data.frame(mydata2),
                "is dataframe?",
                shiny = FALSE),
    silent = TRUE)

#> [SPS-ERROR] 2020-08-10 09:51:58 is dataframe? failed