Dynamically generate Shiny file selection component

Depending on the "mode" in SPS options, this function renders a similar UI components but behaves differently on server. local mode will not copy file, directly use a path pointer, server mode upload file and store in temp. Expect similar behavior as shiny::fileInput.

dynamicFile(
  id,
  title = "Select your file:",
  label = "Browse",
  icon = NULL,
  style = "",
  multiple = FALSE
)

Arguments

id
element ID

title
element title

label
upload button label

icon
button icon, only works for local mode

style
additional button style, only works for local mode

multiple
multiple files allowed

Value

a Shiny upload component

Examples

if(interactive()){
    library(shinyjs)
    # change to 'local' to see the difference
    options(sps = list(mode='server'))
    ui <- fluidPage(
        useShinyjs(),
        dynamicFile("getFile"),
        textOutput("txt_file")
    )

    server <- function(input,output,session){
        runjs('$(".sps-file input").attr("readonly", true)')
        myfile <- dynamicFileServer(input,session, id = "getFile")
        observe({
            print(myfile())
        })
    }
    shinyApp(ui = ui, server = server)
}