Load tabular files as tibbles to server

load a file to server end. It's designed to be used with the input file source switch button(see it in a SPS new template data tab). It uses vroom::vroom to load the file. In SPS, this function is usually combined as downstream of dynamicFileServer() function on on the server side to read the file into R. This loading function only works for parsing tabular data, use vroom::vroom() internally.

loadDF(
  choice,
  data_init = NULL,
  upload_path = NULL,
  eg_path = NULL,
  comment = "#",
  delim = "\t",
  col_types = vroom::cols(),
  ...
)

Arguments

choice
where this file comes from, from 'upload' or example 'eg'?

data_init
a tibble to return if upload_path or eg_path is not provided. Return a 8x8 empty tibble if not provided

upload_path
when choice is "upload", where to load the file, will return data_init if this param is not provided

eg_path
when choice is "eg", where to load the file, will return data_init if this param is not provided

comment
comment characters when load the file, see help file of vroom

delim
delimiter characters when load the file, see help file of vroom

col_types
columns specifications, see help file of vroom

...
other params for vroom, see help file of vroom

Value

returns a tibble or NULL if parsing is unsuccessful

Details

This is function is wrapped by the shinyCatch() function, so it will show loading information both on console and on UI. This function prevents errors to crash the Shiny app, so any kind of file upload will not crash the app. To show message on UI, useSps() must be used in Shiny UI function, see examples.

Examples

if(interactive()){
    library(shinyWidgets)
    # change value to 'local' to see the difference
    spsOption("mode", value = "server")
    ui <- fluidPage(
        useSps(),
        shinyWidgets::radioGroupButtons(
            inputId = "data_source", label = "Choose your data file source:",
            selected = "upload",
            choiceNames = c("Upload", "Example"),
            choiceValues = c("upload", "eg")
        ),
        dynamicFile("data_path", label = "input file"),
        dataTableOutput("df")
    )

    server <- function(input, output, session) {
        tmp_file <- tempfile(fileext = ".csv")
        write.csv(iris, file = tmp_file)
        upload_path <- dynamicFileServer(input, session, "data_path")
        data_df <- reactive({
            loadDF(choice = input$data_source,
                   upload_path = upload_path()$datapath,
                   delim = ",", eg_path = tmp_file)
        })
        output$df <- renderDataTable(data_df())
    }
    shinyApp(ui, server)
}