Create a new SPS Tab

functions to create a new SPS tab. It is recommended to create the data tab first then linked plot tabs.

newTabPlot(
  tab_id = "plot_id1",
  tab_displayname = "Plot Tab Title",
  desc = "default",
  img = "",
  plot_expr = plotly::ggplotly(ggplot(mydata$data, aes_string(names(mydata$data)[1],
    names(mydata$data)[2])) + geom_point(aes(color = seq_len(nrow(mydata$data))))),
  pkgs = list(cran_pkg = c("base"), bioc_pkg = c(""), github = c("")),
  plot_data = list(makePlotData(app_path = app_path)),
  plot_out_func = plotly::plotlyOutput,
  plot_render_func = plotly::renderPlotly,
  app_path = getwd(),
  out_folder_path = file.path(app_path, "R"),
  plot_control_ui = tagList(h3("Some plotting options")),
  author = "",
  empty = FALSE,
  preview = FALSE,
  use_string = FALSE,
  reformat = TRUE,
  open_file = TRUE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)

newTabData(
  tab_id = "data_id1",
  tab_displayname = "Data Tab Title",
  desc = "default",
  pkgs = list(cran_pkg = c("base"), bioc_pkg = c(""), github = c("")),
  common_validation = spsValidate({     "pass" }, "common"),
  prepro_methods = list(makePrepro("nothing", "do nothing"), makePrepro("md1",
    "method1", vd_expr = {     nrow(data_filtered) > 1 })),
  app_path = getwd(),
  out_folder_path = file.path(app_path, "R"),
  eg_path = file.path(app_path, "data", "iris.csv"),
  author = "",
  empty = FALSE,
  preview = FALSE,
  reformat = TRUE,
  open_file = TRUE,
  use_string = FALSE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)

Arguments

tab_id
character string, length 1, must start with "plot_" for plot tabs and "data_" for data tabs. Must be a unique value.

tab_displayname
character string, length 1, the name to be displayed on sidebar tab list and tab

desc
character string, length 1 in markdown format. Tab description and instructions. You can make type it in multiple lines but in only one string (one pair of quotes). e.g.

"
# some desc
## second line,
- bullet 1
- bullet 2
"

img
an example image of the kind of plot users can make out of this tab. It can be a internet link or a local link which uses the www folder as the root. e.g. drop your image plot.png inside www/plot_list, then the link here is "plot_list/plot.png". Only needed for plot tabs.

plot_expr
the plot expression, like all other expression in other shiny reactive expressions. e.g for more than one line use . default is

plotly::ggplotly(
    ggplot(mydata$data,
           aes_string(names(mydata$data)[1],
                      names(mydata$data)[2])) +
        geom_point(aes(
            color = seq_len(nrow(mydata$data))
        ))
)

pkgs
which packages you require users to install, list. specify CRAN, bioconductor or github packages in a vector. see shinyCheckPkg

list(
    cran_pkg = c("base"),
    bioc_pkg = c(""),
    github = c("")
)

plot_data
a list of makePlotData() results, see makePlotData for details.

list(
    makePlotData("plot_1", ...),
    makePlotData("plot_2", ...),
    ...
)

plot_out_func
the plot output function to use on UI, like shiny::plotOutput, just the function name without quotes and without (). default is plotly::plotlyOutput.

plot_render_func
The render plot function to use on server. No quotes, no (). Must be paired with plot_out_func. e.g If plotly::plotlyOutput is used on UI, server must use plotly::renderPlotly. If you use shiny::renderPlot, plot will not show up.

app_path
string, app directory, default is current directory

out_folder_path
string, which directory to write the new tab file, default is the R folder a SPS project

plot_control_ui
additional UI components you want to add to control plotting options, like additional slider bar, some on/off switches, text input etc. If more than one components, put them in a shiny::tagList

author
character string, or a vector of strings. authors of the tab

empty
bool, for advanced developers, if you don't want to use SPS default tab UI and server structure, you can use turn this to TRUE. A very simple template will be generated and you need to write all UI and server parts by yourself. In this case, only tab_id, display_name, author are needed, other tab args can be ignored, system args are still working, like verbose, preview, style, colorful

preview
bool, TRUE will print the new tab code to console and will not write the file and will not register the tab

use_string
bool, sometimes parsing an expression in R may not be totally accurate. To aviod this problem, turn this to TRUE and for plot_control_ui, plot_expr, p_out_func, p_render_func, wrap your expression in a quoted string. What you have provided in the string will be what on the new tab file, no expression parsing will happen. Can only be controlled as a group, which means use string for all of them or none of them in plot tabs. For data tab, the affected argument is common_validation . When turn this to TRUE, be careful with quotes in your expression, escape or use alternative of single/double quotes.

newTabPlot(
    ...
    use_string = TRUE,
    plot_control_ui = "
    tagList(clearableTextInput('id1', 'label')), h5('this title')
    ",
    plot_expr = "
    plotly::ggplotly(
        ggplot(mydata$data,
               aes_string(names(mydata$data)[1],
                          names(mydata$data)[2])) +
            geom_point(aes(
                color = seq_len(nrow(mydata$data))
            ))
    )
    ",
    p_out_func = "plotly::plotlyOutput"
    ...
)

reformat
bool, whether to use styler::style_file reformat the code

open_file
bool, if Rstudio is detected, open the new tab file?

verbose
bool, default follows the project verbosity level. TRUE will give you more information on progress and debugging

colorful
bool, whether the message will be colorful or not

common_validation
expression, use '' to wrap around multiple line expressions. Usually a spsValidate object. You can use shiny's built in shiny::req or shiny::validate for a simpler version.

prepro_methods
a list of makePrepro method returns, read help for that function for details

eg_path
example data set path. Each data tab requires an example data set to be displayed when users don't have anything to upload. Usually this data file is a tabular file and stored in the data folder in a SPS project

Value

a tab file in R folder and tab info registered on config/tabs.csv

Details

Must use this function inside a SPS project, use spsInit() if there is no project.

For a new data tab, different preprocessing methods, their pre-prequirements and what plotting options available after each preprocess is controlled by the makePrepro function. Each call from this function specify one preprocessing method. All preprocess methods should be provided in a list to the prepro_methods argument.

A new plot tab can have more than one data set as the input. For example a plot can require a metadata table and a log transformed table as inputs. There maybe multiple data tabs can preprocess and produce the same log table. So you need to specify how many data inputs this plot requires; for each input which data tab(s) this plot tab can receive data from; for each input data type, what validations (data format checks) you want to do. All of these are controlled by makePlotData and return(s) of this function should be provided in a list to the plot_data argument. is controlled by the makePrepro function. Each call from this function specify one preprocessing method. All preprocess methods should be provided in a list to the prepro_methods argument.

One step in creating plot tab is specify incoming data source options by the makePlotData method. It requires the data tab IDs exist in the config file config/tabs.csv. So, it is best to create all required data tabs first. Or specify it to any existing data tab like 'data_raw' and when the template is created, manually change it.

Examples

spsInit(change_wd = FALSE, overwrite = TRUE)

#> [SPS-INFO] 2020-09-03 00:48:07 Start to create a new SPS project

#> Warning: [SPS-WARNING] 2020-09-03 00:48:07 Overwrite things in /home/lz/Desktop/systemPipeShiny/docs/reference/SPS_20200903

#> [SPS-INFO] 2020-09-03 00:48:07 Now copy files
#> [SPS-INFO] 2020-09-03 00:48:07 Create SPS database
#> [SPS-INFO] 2020-09-03 00:48:07 Created SPS database method container
#> [SPS-INFO] 2020-09-03 00:48:07 Creating SPS db...
#> [SPS-DANGER] 2020-09-03 00:48:07 Db created at '/home/lz/Desktop/systemPipeShiny/docs/reference/SPS_20200903/config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-09-03 00:48:07 Key md5 199397fa9b95bd65100d039652f09c51
#> [SPS-INFO] 2020-09-03 00:48:07 SPS project setup done!

newTabData(
    tab_id = "data_new",
    tab_displayname = "my first data tab",
    prepro_methods = list(makePrepro(label = "do nothing",
                                     plot_options = "plot_new")),
    app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}")
)

#> [SPS-INFO] 2020-09-03 00:48:07 Write to file SPS_20200903/R/tab_vs_data_new.R
#> Styling  1  files:
#>  SPS_20200903/R/tab_vs_data_new.R ℹ 
#> ────────────────────────────────────────
#> Status	Count	Legend 
#> ✓ 	0	File unchanged.
#> ℹ 	1	File changed.
#> x 	0	Styling threw an error.
#> ────────────────────────────────────────
#> [SPS-INFO] 2020-09-03 00:48:10 Now register your new tab to config/tab.csv
#> [SPS-INFO] 2020-09-03 00:48:10 New tab created!

newTabPlot(
    tab_id = "plot_new1",
    tab_displayname = "my first plot tab",
    plot_data = list(
        makePlotData(dataset_label = "Data from my new tab",
                     receive_datatab_ids = "data_new",
                     app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}"))
    ),
    app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}")
)

#> [SPS-INFO] 2020-09-03 00:48:10 Write to file SPS_20200903/R/tab_vs_plot_new1.R
#> Styling  1  files:
#>  SPS_20200903/R/tab_vs_plot_new1.R ℹ 
#> ────────────────────────────────────────
#> Status	Count	Legend 
#> ✓ 	0	File unchanged.
#> ℹ 	1	File changed.
#> x 	0	Styling threw an error.
#> ────────────────────────────────────────
#> [SPS-INFO] 2020-09-03 00:48:12 Now register your new tab to config/tab.csv
#> [SPS-INFO] 2020-09-03 00:48:12 New tab created!

newTabData(
    tab_id = "data_empty",
    tab_displayname = "my first empty data tab",
    empty = TRUE,
    app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}")
)

#> [SPS-INFO] 2020-09-03 00:48:12 Write to file SPS_20200903/R/tab_vs_data_empty.R
#> Styling  1  files:
#>  SPS_20200903/R/tab_vs_data_empty.R ℹ 
#> ────────────────────────────────────────
#> Status	Count	Legend 
#> ✓ 	0	File unchanged.
#> ℹ 	1	File changed.
#> x 	0	Styling threw an error.
#> ────────────────────────────────────────
#> [SPS-INFO] 2020-09-03 00:48:12 Now register your new tab to config/tab.csv
#> [SPS-INFO] 2020-09-03 00:48:12 New tab created!

newTabPlot(
    tab_id = "plot_empty",
    tab_displayname = "my first empty plot tab",
    empty = TRUE,
    app_path = glue("SPS_{format(Sys.time(), '%Y%m%d')}")
)

#> [SPS-INFO] 2020-09-03 00:48:12 Write to file SPS_20200903/R/tab_vs_plot_empty.R
#> Styling  1  files:
#>  SPS_20200903/R/tab_vs_plot_empty.R ℹ 
#> ────────────────────────────────────────
#> Status	Count	Legend 
#> ✓ 	0	File unchanged.
#> ℹ 	1	File changed.
#> x 	0	Styling threw an error.
#> ────────────────────────────────────────
#> [SPS-INFO] 2020-09-03 00:48:13 Now register your new tab to config/tab.csv
#> [SPS-INFO] 2020-09-03 00:48:13 New tab created!