Remove a SPS tab

Remove a tab R file and remove from the tabs.csv config file

removeSpsTab(
  tab_id = "none",
  force = FALSE,
  app_path = getwd(),
  multiple = FALSE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)

Arguments

tab_id
tab ID, string, length 1, supports regular expressions, so be careful. If more than one tabs are matched, stop by default

force
bool, whether to ask for confirmation

app_path
app directory

multiple
bool, if matched more than one tab, turn this to TRUE can remove more than one tab at a time. Be careful.

verbose
bool, follows project setting, but can be overwrite. TRUE will give you more information

colorful
bool, whether the message will be colorful?

Value

remove the tab file and register info in tabs.csv

Examples

spsInit(change_wd = FALSE, overwrite = TRUE)

#> [SPS-INFO] 2020-10-07 21:55:53 Start to create a new SPS project

#> Warning: [SPS-WARNING] 2020-10-07 21:55:53 Overwrite things in /home/lz/Desktop/master/systemPipeShiny/docs/reference/SPS_20201007

#> [SPS-INFO] 2020-10-07 21:55:53 Now copy files

#> [SPS-INFO] 2020-10-07 21:55:53 Create SPS database

#> [SPS-INFO] 2020-10-07 21:55:53 Created SPS database method container

#> [SPS-INFO] 2020-10-07 21:55:53 Creating SPS db...

#> [SPS-DANGER] 2020-10-07 21:55:53 Db created at '/home/lz/Desktop/master/systemPipeShiny/docs/reference/SPS_20201007/config/sps.db'. DO NOT share this file with others
#> [SPS-INFO] 2020-10-07 21:55:53 Key md5 e5d3b59f235239bf6b9f35ee96f361b4
#> [SPS-INFO] 2020-10-07 21:55:53 SPS project setup done!

newTabData(
    tab_id = "data_new_remove",
    tab_displayname = "my first data tab",
    prepro_methods = list(makePrepro(label = "do nothing",
                                     plot_options = "plot_new")),
    app_path = glue::glue("SPS_{format(Sys.time(), '%Y%m%d')}")
)

#> [SPS-INFO] 2020-10-07 21:55:53 Write to file SPS_20201007/R/tab_vs_data_new_remove.R

#> Styling  1  files:
#>  /home/lz/Desktop/master/systemPipeShiny/docs/reference/SPS_20201007/R/tab_vs_data_new_remove.R ℹ 
#> ────────────────────────────────────────────────────────────────────────────────
#> Status	Count	Legend 
#> ✓ 	0	File unchanged.
#> ℹ 	1	File changed.
#> x 	0	Styling threw an error.
#> ────────────────────────────────────────────────────────────────────────────────

#> [SPS-INFO] 2020-10-07 21:55:56 Now register your new tab to config/tab.csv

#> [SPS-SUCCESS] 2020-10-07 21:55:56 New tab created!

removeSpsTab("data_new_remove", force = TRUE,
             app_path = glue::glue("SPS_{format(Sys.time(), '%Y%m%d')}"))

#> [SPS-INFO] 2020-10-07 21:55:56 Matched tab(s): data_new_remove

#> [SPS-INFO] 2020-10-07 21:55:56 1 tabs removed from tabs.csv

#> [SPS-INFO] 2020-10-07 21:55:56 Now remove file SPS_20201007/R/tab_vs_data_new_remove.R

#> [SPS-SUCCESS] 2020-10-07 21:55:56 Removal complete